Migrating test from XCTest to Testing
