+++ Tests/SwiftLayoutTests/Util/Extension/Attribute+CaseIterable.swift
@@ -5,7 +5,7 @@
 import UIKit
 import SwiftLayout
 
-extension NSLayoutConstraint.Attribute: CaseIterable {
+extension NSLayoutConstraint.Attribute: @retroactive CaseIterable {
     public static var allCases: [NSLayoutConstraint.Attribute] {
         (0...20).compactMap {
             NSLayoutConstraint.Attribute(rawValue: $0)
@@ -13,19 +13,19 @@ extension NSLayoutConstraint.Attribute: CaseIterable {
     }
 }
 
-extension AnchorsXAxisAttribute: CaseIterable {
+extension AnchorsXAxisAttribute: @retroactive CaseIterable {
     public static var allCases: [AnchorsXAxisAttribute] {
         NSLayoutConstraint.Attribute.allCases.compactMap { AnchorsXAxisAttribute(attribute: $0) }
     }
 }
 
-extension AnchorsYAxisAttribute: CaseIterable {
+extension AnchorsYAxisAttribute: @retroactive CaseIterable {
     public static var allCases: [AnchorsYAxisAttribute] {
         NSLayoutConstraint.Attribute.allCases.compactMap { AnchorsYAxisAttribute(attribute: $0) }
     }
 }
 
-extension AnchorsDimensionAttribute: CaseIterable {
+extension AnchorsDimensionAttribute: @retroactive CaseIterable {
     public static var allCases: [AnchorsDimensionAttribute] {
         NSLayoutConstraint.Attribute.allCases.compactMap { AnchorsDimensionAttribute(attribute: $0) }
     }
